pyproject.toml
[project]
name = "research-mcp"
version = "0.1.0"
description = "Deep research MCP server."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.1",
    "strands-agents[ollama]>=1.5.0",
]

[project.scripts]
research-mcp-cli = "cli.main:main_sync"
research-mcp = "mcp_server.server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["mypy>=1.17.1", "pytest>=8.4.1", "ruff>=0.12.9"]

[tool.hatch.build.targets.wheel]
packages = ["src/cli"]

[tool.ruff]
# Formatting and linting configuration
line-length = 88
indent-width = 4

# Assume Python 3.12+.
target-version = "py312"

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
# Also enable import sorting (`I`), unused imports (`F401`), and more!
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings  
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PTH", # flake8-use-pathlib
]
ignore = [
    "E501", # line too long, handled by ruff formatter
    "W191", # indentation contains tabs
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces, not tabs
indent-style = "space"

# Respect magic trailing commas
skip-magic-trailing-comma = false

# Auto-detect line endings
line-ending = "auto"

[tool.pytest.ini_options]
# Add src directory to Python path for imports
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
# Add src directory to Python path for mypy too
mypy_path = "src"
# Use explicit package bases to avoid module name conflicts
explicit_package_bases = true
# Show error codes
show_error_codes = true
# Skip checking external libraries without stubs
ignore_missing_imports = true
